[Fix]migrate

+++ db/schema.rb
@@ -59,8 +59,8 @@ ActiveRecord::Schema.define(version: 2023_08_12_141359) do
   end
 
   create_table "comments", force: :cascade do |t|
-    t.integer "post_id", null: false
-    t.integer "user_id", null: false
+    t.bigint "post_id", null: false
+    t.bigint "user_id", null: false
     t.text "comment", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
@@ -69,8 +69,8 @@ ActiveRecord::Schema.define(version: 2023_08_12_141359) do
   end
 
   create_table "favorites", force: :cascade do |t|
-    t.integer "post_id", null: false
-    t.integer "user_id", null: false
+    t.bigint "post_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["post_id"], name: "index_favorites_on_post_id"
@@ -78,8 +78,8 @@ ActiveRecord::Schema.define(version: 2023_08_12_141359) do
   end
 
   create_table "posts", force: :cascade do |t|
-    t.integer "user_id", null: false
-    t.integer "category_id", null: false
+    t.bigint "user_id", null: false
+    t.bigint "category_id", null: false
     t.string "spot_name", null: false
     t.string "title", null: false
     t.text "comment", null: false
